backend/requirements.txt
fastapi>=0.135
uvicorn>=0.41
psycopg2-binary>=2.9
redis>=7.3
httpx>=0.28
python-jose[cryptography]>=3.5
pydantic>=2.12
